Clean up, merge files and start test
